learning-gem5: Add a simple cpu model for Part IV
The idea is to make a simple CPU model so I can understand how the CPU
models in gem5 work. This is part of the Learning gem5 book. See
learning.gem5.org/ for more information.

This simple CPU is similar in performance to SimpleTimingCPU. It is a
"multi-cycle" design, but execution takes one cycle for all instructions
except memory instructions.

Included in the code is a MemoryRequest type that wraps all memory
requests and the required ExecContext which is a simple wrapper around
SimpleThread.

In the future, it would be best to replace SimpleThread with a better
documented (hopefully simpler) class.

The code currently only supports SE mode as interrupts are not correctly
enabled.

This has been tested with both X86 and RISC-V. Testing with ARM is furture
work.

Change-Id: I0d33dd7c8174be8ab2d7f359d2baf5e9ab2740b3
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>