sim, util: Change how shared memory is used in external hypercall
This commit changes how shared memory is used when a hypercall
is triggered by an external process. Unlinking the shared
memory has been moved to the transmitter side, and the signal
handler in gem5 writes a "done" message to the shared memory to
let the transmitter know to unlink it. The shared memory name
has also been changed to include the gem5 process's PID, so
separate gem5 processes won't have the same shared memory.